port is type int, not short, cast discoveryPort to int.
git-svn-id: https://svn.apache.org/repos/asf/river/jtsk/trunk@1741922 13f79535-47bb-0310-9956-ffa450edef68

+++ src/net/jini/core/discovery/LookupLocator.java
@@ -408,7 +408,7 @@ public class LookupLocator implements Serializable {
         GetField fields = oin.readFields();
 	host = (String) fields.get("host", null);
 	if (host == null) throw new InvalidObjectException("host cannot be null");
-	port = fields.get("port", discoveryPort);
+	port = fields.get("port", (int) discoveryPort);
     }
     
     /**
